researchpapersummizer_backend/requirements.txt

django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
openai
requests

pypdf==3.17.1
python-dotenv==1.0.0

google-generativeai

gunicorn
whitenoise
dj-database-url
psycopg2-binary
